pyproject.toml
[tool.poetry]
name = "messaging-system"
version = "0.1.0"
description = "Exchange messages between users"
authors = ["Leonid"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
python-dotenv = "^0.13.0"
passlib = "^1.7.2"
bcrypt = "^3.1.7"
gunicorn = "^20.0.4"
flask-jwt-extended = "^3.24.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^19.10b0", allow-prereleases = true}
flake8 = "^3.8.2"

[tool.black]
line-length = 100
target-version = ['py38']
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
  )/
)
'''

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
